Move `Assigned<F>` into a separate module
This doesn't alter the public crate API; it's just to make it easier to
manage the struct implementation.
